Added the ability to masquerade as a specific account.
I added an option for a developer config file in I3.config.  If the developer config file has a
key/value pair named "masquerade_as_account" with the account name as a value, then
I3.server.remote_account_as will return that value instead of the value passed in the HTTP headers.
This allows developers to test how the Intranet behaves as different users without having to know
their password.

If the "developer.yml" file is not present, or the "masquerade_as_account" key is not present, the
normal (default) behavior will occur.